refactor(core): move state ownership from Harness v1 to Session
Session now owns state machinery (getState, setState, updateState,
schema validation, update queue). Harness only passes stateSchema and
initialState config to created sessions. HarnessCompat composes legacy
harness state with session-owned state and session-derived fields
(currentModelId, modeId).
