+++ packages/core/src/harness/v1/harness.state-workspace.test.ts
@@ -64,8 +64,8 @@ const createHarness = <TState extends Record<string, unknown>>(config: TestHarne
   return { harness, storage, memory };
 };
 
-describe('Harness v1 state', () => {
-  it('initializes from schema defaults and initial state', () => {
+describe('Harness v1 session state', () => {
+  it('initializes sessions from schema defaults and initial state', async () => {
     const { harness } = createHarness<{ count: number; label: string }>({
       stateSchema: {
         type: 'object',
@@ -75,8 +75,10 @@ describe('Harness v1 state', () => {
       initialState: { count: 2, label: 'ready' },
     });
 
-    expect(harness.getState()).toEqual({ count: 2, label: 'ready' });
-    expect(Object.isFrozen(harness.getState())).toBe(true);
+    const session = await harness.session({ threadId: 'thread-1', resourceId: 'resource-1' });
+
+    expect(session.getState()).toEqual({ count: 2, label: 'ready' });
+    expect(Object.isFrozen(session.getState())).toBe(true);
   });
 
   it('validates setState and emits state_changed events', async () => {
@@ -92,28 +94,30 @@ describe('Harness v1 state', () => {
       events.push(event);
     });
 
-    await harness.setState({ count: 1 });
+    const session = await harness.session({ threadId: 'thread-1', resourceId: 'resource-1' });
+    await session.setState({ count: 1 });
 
-    expect(harness.getState()).toEqual({ count: 1 });
+    expect(session.getState()).toEqual({ count: 1 });
     expect(events).toContainEqual(
       expect.objectContaining({ type: 'state_changed', state: { count: 1 }, changedKeys: ['count'] }),
     );
-    await expect(harness.setState({ count: 'bad' as never })).rejects.toThrow('Invalid state update');
-    expect(harness.getState()).toEqual({ count: 1 });
+    await expect(session.setState({ count: 'bad' as never })).rejects.toThrow('Invalid state update');
+    expect(session.getState()).toEqual({ count: 1 });
   });
 
   it('runs updateState as a serialized transaction', async () => {
     const { harness } = createHarness<{ count: number }>({
       initialState: { count: 0 },
     });
+    const session = await harness.session({ threadId: 'thread-1', resourceId: 'resource-1' });
 
-    const result = await harness.updateState(state => ({
+    const result = await session.updateState(state => ({
       updates: { count: state.count + 1 },
       result: state.count,
     }));
 
     expect(result).toBe(0);
-    expect(harness.getState()).toEqual({ count: 1 });
+    expect(session.getState()).toEqual({ count: 1 });
   });
 });
